Fix most warnings
* remove not needed gitignore from /src, it causes a duplicate resource
warning for the binary path (this could also be fixed by enabling a
resource copy filter in the JDT project properties)
* remove unused imports
* disable discouraged access warnings in Java
* disable primitive valued warnings in Xtend

There are no functional changes, and the only remaining warnings are
some grammar warnings. The disablement of warnings sets single project
properties and does not contain complete .prefs files to modify as
little as possible.