ignored plugin has to exactly match the version, if any - #95
When a plugin to be ignored is declared with a version number, but that
version does not actually exist at all, we have ignored the most recent
version available for that plugin name.
Instead, this 'ignore' statement should not apply, because ignoring a
specific version usually has a specific reason related to that version.